[datafeed] Move all response saving/loading into FRCAPI class
