+++ src/backend/tasks_io/datafeeds/tests/datafeed_fms_api_test.py
@@ -11,15 +11,8 @@ from backend.common.sitevars.fms_api_secrets import (
     ContentType as FMSApiSecretsContentType,
 )
 from backend.common.sitevars.fms_api_secrets import FMSApiSecrets
-from backend.common.storage.clients.in_memory_client import InMemoryClient
 from backend.common.urlfetch import URLFetchResult
 from backend.tasks_io.datafeeds.datafeed_fms_api import DatafeedFMSAPI
-
-
-@pytest.fixture(autouse=True)
-def reset_gcs_client():
-    yield
-    InMemoryClient.CLIENT = None
 
 
 @pytest.fixture()
@@ -31,7 +24,7 @@ def test_init(ndb_stub):
     with pytest.raises(
         Exception, match="Missing FRC API auth token. Setup fmsapi.secrets sitevar."
     ):
-        DatafeedFMSAPI(save_response=True)
+        DatafeedFMSAPI()
 
 
 @pytest.mark.parametrize(
@@ -93,7 +86,7 @@ def test_get_root(fms_api_secrets):
         json.dumps(content),
     )
 
-    df = DatafeedFMSAPI(save_response=True)
+    df = DatafeedFMSAPI()
     with patch.object(
         FRCAPI, "root", return_value=InstantFuture(response)
     ) as mock_root:
@@ -110,7 +103,7 @@ def test_get_root_failure(fms_api_secrets):
         json.dumps(content),
     )
 
-    df = DatafeedFMSAPI(save_response=True)
+    df = DatafeedFMSAPI()
     with patch.object(
         FRCAPI, "root", return_value=InstantFuture(response)
     ) as mock_root:
@@ -132,7 +125,7 @@ def test_mark_api_down(fms_api_secrets):
         "{}",
     )
 
-    df = DatafeedFMSAPI(save_response=True)
+    df = DatafeedFMSAPI()
     with patch.object(FRCAPI, "root", return_value=InstantFuture(response1)):
         assert df.get_root_info().get_result() is None
         assert ApiStatusFMSApiDown.get() is True
@@ -140,115 +133,3 @@ def test_mark_api_down(fms_api_secrets):
     with patch.object(FRCAPI, "root", return_value=InstantFuture(response2)):
         assert df.get_root_info().get_result() == {}
         assert ApiStatusFMSApiDown.get() is False
-
-
-def test_save_response(fms_api_secrets, monkeypatch: pytest.MonkeyPatch):
-    monkeypatch.setenv("SAVE_FRC_API_RESPONSE", "true")
-    content = {
-        "currentSeason": 2021,
-        "maxSeason": 2021,
-        "name": "FIRST ROBOTICS COMPETITION API",
-        "apiVersion": "3.0",
-        "status": "normal",
-    }
-    response = URLFetchResult.mock_for_content(
-        "https://frc-api.firstinspires.org/v3.0/root",
-        200,
-        json.dumps(content),
-    )
-
-    df = DatafeedFMSAPI(save_response=True)
-    with patch.object(FRCAPI, "root", return_value=InstantFuture(response)):
-        df.get_root_info().get_result()
-
-    client = InMemoryClient.get()
-    files = client.get_files()
-    assert len(files) == 1
-
-    f = client.read(files[0])
-    assert f is not None
-    assert f == response.content
-
-
-def test_save_response_unchanged(fms_api_secrets, monkeypatch: pytest.MonkeyPatch):
-    monkeypatch.setenv("SAVE_FRC_API_RESPONSE", "true")
-    content = {
-        "currentSeason": 2021,
-        "maxSeason": 2021,
-        "name": "FIRST ROBOTICS COMPETITION API",
-        "apiVersion": "3.0",
-        "status": "normal",
-    }
-
-    response = URLFetchResult.mock_for_content(
-        "https://frc-api.firstinspires.org/v3.0/root",
-        200,
-        json.dumps(content),
-    )
-
-    df = DatafeedFMSAPI(save_response=True)
-    with patch.object(FRCAPI, "root", return_value=InstantFuture(response)):
-        df.get_root_info().get_result()
-
-    client = InMemoryClient.get()
-    files = client.get_files()
-    assert len(files) == 1
-    f_name = files[0]
-
-    with patch.object(FRCAPI, "root", return_value=InstantFuture(response)):
-        df.get_root_info().get_result()
-
-    # Since the content didn't change, we shouldn't have written another
-    assert client.get_files() == [f_name]
-
-
-def test_save_response_updated(fms_api_secrets, monkeypatch: pytest.MonkeyPatch):
-    monkeypatch.setenv("SAVE_FRC_API_RESPONSE", "true")
-    content = {
-        "currentSeason": 2021,
-        "maxSeason": 2021,
-        "name": "FIRST ROBOTICS COMPETITION API",
-        "apiVersion": "3.0",
-        "status": "normal",
-    }
-
-    response = URLFetchResult.mock_for_content(
-        "https://frc-api.firstinspires.org/v3.0/root",
-        200,
-        json.dumps(content),
-    )
-
-    df = DatafeedFMSAPI(save_response=True)
-    with patch.object(FRCAPI, "root", return_value=InstantFuture(response)):
-        df.get_root_info().get_result()
-
-    client = InMemoryClient.get()
-    files = client.get_files()
-    assert len(files) == 1
-
-    content2 = {
-        "currentSeason": 2021,
-        "maxSeason": 2021,
-        "name": "SECOND ROBOTICS COMPETITION API",
-        "apiVersion": "3.0",
-        "status": "normal",
-    }
-    response2 = URLFetchResult.mock_for_content(
-        "https://frc-api.firstinspires.org/v3.0/root",
-        200,
-        json.dumps(content2),
-    )
-    with patch.object(FRCAPI, "root", return_value=InstantFuture(response2)):
-        df.get_root_info().get_result()
-
-    # Since the content is different, we should have two items
-    files = client.get_files()
-    assert len(files) == 2
-
-    f = client.read(files[0])
-    assert f is not None
-    assert f == response.content
-
-    f2 = client.read(files[1])
-    assert f2 is not None
-    assert f2 == response2.content
